Fix wrong order deletion
On windows the order of deletions is not the same than on Linux.
On Linux it seems like thread local variables are deleted BEFORE than
static variables.
On Linux the thread local variables are deleted after (it seems). So we
need to enforce that the thread's buffer are not deleted when the buffer
set will be.