Simplify the memory allocation trace info.
Don't use two different containers anymore for allocations and events,
just increase the size of the event value and remove the thread info
from every event. The thread info is in the header, so, we can use that
from there directly in the parser and don't duplicate information (bytes
don't grow in trees).